requirements.txt
pyserial
sv_ttk
Pillow
ttkbootstrap
